# Add glob support to ignores