--- a/config.go
+++ b/config.go
@@ -145,43 +145,7 @@ func (config *Config) getIndex(list ls) (index, error) {
 	return index, nil
 }
 
-func (config *Config) globIndex(filesystem fs.FS, index index) (err error) {
-	for key, value := range index {
-		var matches []string
-
-		if !strings.ContainsAny(key, "*{}") {
-			continue
-		}
-
-		if matches, err = doublestar.Glob(filesystem, key); err != nil {
-			return err
-		}
-
-		if len(matches) == 0 {
-			delete(index, key)
-			continue
-		}
-
-		for _, match := range matches {
-			var matchInfo fs.FileInfo
-
-			if matchInfo, err = fs.Stat(filesystem, match); err != nil {
-				return err
-			}
-
-			if !matchInfo.IsDir() {
-				continue
-			}
-
-			index[match] = value
-			delete(index, key)
-		}
-	}
-
-	return nil
-}
-
-func globIgnoreIndex[V bool | map[string][]Rule](config *Config, filesystem fs.FS, index map[string]V) (err error) {
+func globIndex[V bool | map[string][]Rule](config *Config, filesystem fs.FS, index map[string]V) (err error) {
 	for key, value := range index {
 		var matches []string