As is on the v2 wishlist and was mentioned in https://github.com/loeffel-io/ls-lint/issues/35 and https://github.com/loeffel-io/ls-lint/issues/56#issuecomment-909418833, globbing is currently not allowed for the `ignore` key.

This is a first try at implementing this.

* [x] Performance implications?
* [ ] Documentation?
* [ ] Haven't found a changelog

### Performance

All measurements done on an 2020 M1 MacBook Air with 8GB of ram.
Using hyperfine, I ran `hyperfine './ls-lint'` with the current local build of ls-lint as well as with master. I did a couple variations, based on the [example config](https://github.com/loeffel-io/ls-lint/blob/master/examples/nuxt-nuxt-js/.ls-lint.yml):

| build | config | result |
| ----- | ------ | ------ |
| this branch | with `node_modules` in ignore | `Time (mean ± σ):      16.0 ms ±   0.8 ms    [User: 7.4 ms, System: 8.7 ms]` |
| this branch | with `'**/node_modules'` in ignore | `Time (mean ± σ):     282.4 ms ±  21.8 ms    [User: 71.7 ms, System: 210.1 ms]` |
| master | with `node_module` in ignore | `Time (mean ± σ):      15.9 ms ±   0.7 ms    [User: 7.3 ms, System: 8.7 ms]` |

This kind of confirms what makes some sense: If you use a glob in ignores, depending on the size of your project, you pay a _significant_ penalty in performance. If you don't use any globs, you don't pay measurably.