@loeffel-io 2022-05-29
Great! Thanks for the PR! I am currently on vacation and get back to you asap