@MagnusHJensen 2022-08-01
What's the status on this? Really needed for monorepo structure