@GeorgeTaveras1231 2022-10-07
Would be great to get this merged and published. There doesn't seem to be a good way to ignore certain folder names in arbitrary locations in a repo. I have a monorepo with multiple __tests__ folders that I would like to ignore.