@loeffel-io 2022-10-07
Hey folks,

sorry that this takes so long - the support will come! This will come together with https://github.com/github/super-linter/issues/2505#issuecomment-1263886322 as soon i have enough mind capacity