@dror-weiss 2023-03-18
ping from here as well; I need glob support in the `ignore` section for my monorepo 🙏🏻