@gbrandt1 2023-04-23
no support for ignore globs is a show-stopper for me.