@loeffel-io 2023-05-21
Very happy to announce that v2.0.0-beta.0 is now available: https://github.com/loeffel-io/ls-lint/releases/tag/v2.0.0-beta.0
Please see the changelog for any additional info: https://ls-lint.org/2.0/prologue/changelog.html#v2-0-0-beta

This feature (add glob support to ignores) is planned for `v2.1`
I will close this PR for now (there will be a new one)