@loeffel-io 2023-07-09
This is now supported with `v2.1.0-beta.0` 🎉  feel free to test it! 
more infos: https://github.com/loeffel-io/ls-lint/pull/119